refactor of project structure
