+++ cmd/api/connection.go
@@ -1,4 +1,4 @@
-package cmd
+package api
 
 import (
 	"encoding/json"
