+++ cmd/api/irail-api_test.go
@@ -1,4 +1,4 @@
-package cmd
+package api
 
 import (
 	"reflect"
